kv_translate: store an uncorrected map at serving precision too
Casting the corrected arms fixed half of an asymmetry and exposed the
other half. A corrected map applied itself at whatever precision it
was stored in, so leaving it in double ran the dominant matrix
multiply there. A plain map does the opposite: it casts to the input
dtype on every call, so leaving it in double converts the entire
weight set on each apply instead of once.

Fixing only the first made the uncorrected arms the slow ones, and a
folded arm appeared to beat the map it had been folded into, which is
impossible. Both are stored at serving precision now, and the folded
arms land exactly on their uncorrected counterparts.

Generated-by: Claude Opus 5 (1M context)

+++ research/kv_translate/run_h1.py
@@ -533,7 +533,18 @@ def main() -> int:
 
     arms, reports = {}, {}
     for name, base in inits.items():
-        arms[name] = {"k": base["k"], "v": base["v"], "corrected": False}
+        # Stored at serving precision, like every other arm. A map left at the
+        # double precision it was solved in is converted on each apply --
+        # AffineMap.apply casts to the input's dtype every call -- so an
+        # uncorrected arm would pay a conversion of its whole weight set that
+        # a corrected arm, whose weights were cast once when it was built,
+        # does not. That asymmetry is the mirror image of the one that made
+        # the corrected arms look slow, and it flatters them just as much.
+        arms[name] = {
+            "k": base["k"].cast(sdtype),
+            "v": base["v"].cast(sdtype),
+            "corrected": False,
+        }
     promoted = [f"unw_k{args.k_best}", f"wtd_k{kb}"]
     for base_name in promoted:
         for linear in (True, False):
